+++ Assets/Chapter 6/Example 6.1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4ce901ffea102824ba8a8839b4ff6f22
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
