+++ Assets/Chapter 6/Example 6.1/Chapter 6 Figure 1.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9152e2a0a951fc04e9f2a6f8af6c39bf
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
