+++ Assets/Chapter 6/Example 6.1/Follower Sphere.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 627292a34a6550242afcd226b16d2bae
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
