+++ Assets/Chapter 6/Example 6.1/Main Camera.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3cf0808428cb3da4c899af710ce26356
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
