+++ Assets/Chapter 6/Example 6.2.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 22807a74c2112664eb9d4d940667ab3b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
