+++ Assets/Chapter 6/Example 6.2/Chapter 6 Figure 2.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 258c8e83a038af143aa482ee8069b523
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
