+++ Assets/Chapter 6/Example 6.2/Chapter6Fig2.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 
 public class Chapter6Fig2 : MonoBehaviour
 {
+    public Camera camera;
     public GameObject vehicle;
     public GameObject target;
 
@@ -11,24 +12,20 @@ public class Chapter6Fig2 : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        vehicle = Instantiate(vehicle);
-        target.transform.position = new Vector3(2f, 1f, -10f);
-
+        Cursor.visible = false;
     }
 
     // Update is called once per frame
-    void FixedUpdate()
+    void Update()
     {
-        float h = Input.GetAxis("Horizontal");
-        float v = Input.GetAxis("Vertical");
-        float speed = 10f;
-
-        Vector3 tempVect = new Vector3(h, v, 0);
-        tempVect = tempVect.normalized * speed * Time.deltaTime;
-
-        target.transform.position += tempVect;
-
+        //set gameObject's position equal to the mouse's;
+        target.transform.position = MousePosition(camera);
         vehicle.GetComponent<vehicleChapter6_2>().arrive(target.transform.position);
+    }
 
+    Vector2 MousePosition(Camera camera)
+    {
+        // Track the Vector2 of the mouse's position
+        return camera.ScreenToWorldPoint(Input.mousePosition);
     }
 }
