+++ Assets/Chapter 6/Example 6.2/Chapter6Fig2.cs.meta
