+++ Assets/Chapter 6/Example 6.2/Follower Sphere 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e704bd4628ede32409c3b43a97b80814
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
