+++ Assets/Chapter 6/Example 6.2/Main Camera 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 78bc2f761ab8d0e46aceadd8b0628465
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
