+++ Assets/Chapter 6/Example 6.2/vehicleChapter6_2.prefab
