+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.2.16f1
-m_EditorVersionWithRevision: 2019.2.16f1 (b9898e2d04a4)
+m_EditorVersion: 2019.3.11f1
+m_EditorVersionWithRevision: 2019.3.11f1 (ceef2d848e70)
